Moved the usersyncers into their own package. Pulled the cookie parsing code into the usersync package.
