+++ adapters/openrtb_util_test.go
@@ -1,10 +1,12 @@
 package adapters
 
 import (
+	"testing"
+
 	"github.com/mxmCherry/openrtb"
 	"github.com/prebid/prebid-server/pbs"
+	"github.com/prebid/prebid-server/usersync"
 	"github.com/stretchr/testify/assert"
-	"testing"
 )
 
 func TestCommonMediaTypes(t *testing.T) {
@@ -370,7 +372,7 @@ func TestOpenRTBEmptyUser(t *testing.T) {
 }
 
 func TestOpenRTBUserWithCookie(t *testing.T) {
-	pbsCookie := pbs.NewPBSCookie()
+	pbsCookie := usersync.NewPBSCookie()
 	pbsCookie.TrySync("test", "abcde")
 	pbReq := pbs.PBSRequest{
 		User: &openrtb.User{},
