+++ pbs/pbsrequest.go
@@ -166,11 +166,11 @@ type PBSRequest struct {
 	SDK           *SDK            `json:"sdk"`
 
 	// internal
-	Bidders []*PBSBidder  `json:"-"`
-	User    *openrtb.User `json:"-"`
-	Cookie  *PBSCookie    `json:"-"`
-	Url     string        `json:"-"`
-	Domain  string        `json:"-"`
+	Bidders []*PBSBidder        `json:"-"`
+	User    *openrtb.User       `json:"-"`
+	Cookie  *usersync.PBSCookie `json:"-"`
+	Url     string              `json:"-"`
+	Domain  string              `json:"-"`
 	Start   time.Time
 }
 
@@ -261,7 +261,7 @@ func ParsePBSRequest(r *http.Request, cache cache.Cache, hostCookieSettings *Hos
 
 	// use client-side data for web requests
 	if pbsReq.App == nil {
-		pbsReq.Cookie = ParsePBSCookieFromRequest(r, &(hostCookieSettings.OptOutCookie))
+		pbsReq.Cookie = usersync.ParsePBSCookieFromRequest(r, &(hostCookieSettings.OptOutCookie))
 
 		// Host has right to leverage private cookie store for user ID
 		if uid, _, _ := pbsReq.Cookie.GetUID(hostCookieSettings.Family); uid == "" && hostCookieSettings.CookieName != "" {
