--- usersync/facebook.go
@@ -1,12 +1,0 @@
-package usersync
-
-func NewFacebookSyncer(syncUrl string) Usersyncer {
-	return &syncer{
-		familyName: "audienceNetwork",
-		syncInfo: &UsersyncInfo{
-			URL:         syncUrl,
-			Type:        "redirect",
-			SupportCORS: false,
-		},
-	}
-}
