--- usersync/indexExchange.go
@@ -1,12 +1,0 @@
-package usersync
-
-func NewIndexSyncer(userSyncURL string) Usersyncer {
-	return &syncer{
-		familyName: "indexExchange",
-		syncInfo: &UsersyncInfo{
-			URL:         userSyncURL,
-			Type:        "redirect",
-			SupportCORS: false,
-		},
-	}
-}
