+++ usersync/usersyncers/adform.go
@@ -1,14 +1,16 @@
-package usersync
+package usersyncers
 
 import (
 	"fmt"
 	"net/url"
+
+	"github.com/prebid/prebid-server/usersync"
 )
 
-func NewAdformSyncer(usersyncURL string, externalURL string) Usersyncer {
+func NewAdformSyncer(usersyncURL string, externalURL string) *syncer {
 	redirectUri := fmt.Sprintf("%s/setuid?bidder=adform&uid=$UID", externalURL)
 
-	info := &UsersyncInfo{
+	info := &usersync.UsersyncInfo{
 		URL:         fmt.Sprintf("%s%s", usersyncURL, url.QueryEscape(redirectUri)),
 		Type:        "redirect",
 		SupportCORS: false,
