+++ usersync/usersyncers/adform_test.go
@@ -1,13 +1,13 @@
-package usersync
+package usersyncers
 
 import (
 	"testing"
 )
 
-func TestAppNexusSyncer(t *testing.T) {
-	an := NewAppnexusSyncer("localhost")
+func TestAdformSyncer(t *testing.T) {
+	an := NewAdformSyncer("//cm.adform.net?return_url=", "localhost")
 	syncInfo := an.GetUsersyncInfo()
-	if syncInfo.URL != "//ib.adnxs.com/getuid?localhost%2Fsetuid%3Fbidder%3Dadnxs%26uid%3D%24UID" {
+	if syncInfo.URL != "//cm.adform.net?return_url=localhost%2Fsetuid%3Fbidder%3Dadform%26uid%3D%24UID" {
 		t.Fatalf("should have matched")
 	}
 	if syncInfo.Type != "redirect" {
