+++ usersync/usersyncers/adtelligent.go
@@ -1,18 +1,20 @@
-package usersync
+package usersyncers
 
 import (
 	"fmt"
 	"net/url"
+
+	"github.com/prebid/prebid-server/usersync"
 )
 
-func NewAdtelligentSyncer(externalURL string) Usersyncer {
+func NewAdtelligentSyncer(externalURL string) *syncer {
 
 	redirectURI := fmt.Sprintf("%s/setuid?bidder=adtelligent&uid={uid}", externalURL)
 	usersyncURL := "//sync.adtelligent.com/csync?t=p&ep=0&redir="
 
 	return &syncer{
 		familyName: "adtelligent",
-		syncInfo: &UsersyncInfo{
+		syncInfo: &usersync.UsersyncInfo{
 			URL:         fmt.Sprintf("%s%s", usersyncURL, url.QueryEscape(redirectURI)),
 			Type:        "redirect",
 			SupportCORS: false,
