+++ usersync/usersyncers/eplanning.go
@@ -1,14 +1,16 @@
-package usersync
+package usersyncers
 
 import (
 	"fmt"
 	"net/url"
+
+	"github.com/prebid/prebid-server/usersync"
 )
 
-func NewEPlanningSyncer(usersyncURL string, externalURL string) Usersyncer {
+func NewEPlanningSyncer(usersyncURL string, externalURL string) *syncer {
 	redirectUri := fmt.Sprintf("%s/setuid?bidder=eplanning&uid=$UID", externalURL)
 
-	info := &UsersyncInfo{
+	info := &usersync.UsersyncInfo{
 		URL:         fmt.Sprintf("%s%s", usersyncURL, url.QueryEscape(redirectUri)),
 		Type:        "redirect",
 		SupportCORS: false,
