+++ usersync/usersyncers/openx.go
@@ -1,18 +1,20 @@
-package usersync
+package usersyncers
 
 import (
 	"fmt"
 	"net/url"
 	"strings"
+
+	"github.com/prebid/prebid-server/usersync"
 )
 
-func NewOpenxSyncer(externalURL string) Usersyncer {
+func NewOpenxSyncer(externalURL string) *syncer {
 	externalURL = strings.TrimRight(externalURL, "/")
 	redirectURL := fmt.Sprintf("%s/setuid?bidder=openx&uid=${UID}", externalURL)
 
 	return &syncer{
 		familyName: "openx",
-		syncInfo: &UsersyncInfo{
+		syncInfo: &usersync.UsersyncInfo{
 			URL:         fmt.Sprintf("https://rtb.openx.net/sync/prebid?r=%s", url.QueryEscape(redirectURL)),
 			Type:        "redirect",
 			SupportCORS: false,
