+++ usersync/usersyncers/pulsepoint.go
@@ -1,16 +1,18 @@
-package usersync
+package usersyncers
 
 import (
 	"fmt"
 	"net/url"
+
+	"github.com/prebid/prebid-server/usersync"
 )
 
-func NewPulsepointSyncer(externalURL string) Usersyncer {
+func NewPulsepointSyncer(externalURL string) *syncer {
 	redirect_uri := fmt.Sprintf("%s/setuid?bidder=pulsepoint&uid=%s", externalURL, "%%VGUID%%")
 	usersyncURL := "//bh.contextweb.com/rtset?pid=561205&ev=1&rurl="
 	return &syncer{
 		familyName: "pulsepoint",
-		syncInfo: &UsersyncInfo{
+		syncInfo: &usersync.UsersyncInfo{
 			URL:         fmt.Sprintf("%s%s", usersyncURL, url.QueryEscape(redirect_uri)),
 			Type:        "redirect",
 			SupportCORS: false,
